% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/late_rest.R
\name{create.score.groups}
\alias{create.score.groups}
\title{Create groups from score prediction}
\usage{
create.score.groups(
  dat,
  treat,
  instrument,
  controls,
  n_groups,
  pred_method = "Causal_Forest",
  n_folds = 2
)
}
\arguments{
\item{dat}{A data.frame containing the necessary variables to run the model.}

\item{treat}{Name of the treatment variable}

\item{instrument}{Name of the instrument}

\item{controls}{Controls. Either as a formula, eg ~ x1 + x2, or a vector of strings, eg c("x1", "x2")}

\item{n_groups}{Number of groups to cut score into}

\item{pred_method}{Score prediction method}

\item{n_folds}{Number of folds to use for score prediction}
}
\value{
data.table
}
\description{
Create groups from score prediction
}
\examples{
# Loading packages
library(fixest)

# Generate a simulated dataset
data <- sim.data()

data2 <- create.score.groups(data, treat = "d", instrument = "z", controls = ~x,
                             n_groups = 10)

# Print summary of predicted compliance scores
summary(data2$pred_p)

# Run a standard IV
reg1 <- feols(data = data2, y ~ 1 | d ~ z)

# Run basic test-and-select method
reg2 <- run.late.rest(dat = data2, yname = "y", treat = "d", instrument = "z", controls = ~g)

# Run basic test-and-select method
reg3 <- run.late.rest(dat = data2, yname = "y", treat = "d", instrument = "z",
                                   controls = ~score_g)

# Comparing both regressions
etable(reg1, reg2, reg3)
}
