% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/late.rest-package.R
\docType{package}
\name{late.rest-package}
\alias{late.rest}
\alias{late.rest-package}
\title{late.rest: Improving LATE estimation by a Test-and-Select procedure}
\description{
Implements the Test-and-Select procedure detailed in Hazard and Lowe (2023).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simon-lowe/late.rest}
  \item Report bugs at \url{https://github.com/simon-lowe/late.rest/issues}
}

}
\keyword{internal}
