% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{rand_n_list_group}
\alias{rand_n_list_group}
\title{Generate an group index list of length n with n_groups groups}
\usage{
rand_n_list_group(n, n_groups)
}
\arguments{
\item{n}{Length of the list}

\item{n_groups}{Number of groups}
}
\value{
An integer vector
}
\description{
Generate an group index list of length n with n_groups groups
}
\examples{
# Group index list with 10 observations and 3 groups
rand_n_list_group(10, 3)
}
