% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/late_rest.R
\name{run.late.rest}
\alias{run.late.rest}
\title{Main function}
\usage{
run.late.rest(
  data,
  yname,
  treat,
  instrument,
  controls,
  n_folds = 2,
  p_th = 0.05
)
}
\arguments{
\item{data}{data.frame}

\item{yname}{dependent variable name}

\item{treat}{bla}

\item{instrument}{blu}

\item{controls}{shu}

\item{n_folds}{sha}

\item{p_th}{Group selection threshold}
}
\value{
tada
}
\description{
Main function
}
\examples{
# You can do the following hihi
}
