% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/late_rest.R
\name{sim.data}
\alias{sim.data}
\title{Simulate data}
\usage{
sim.data(
  nsims = 1,
  beta = 1,
  alpha = 0.5,
  n = 1000,
  J = 10,
  rde = 0.3,
  sd = 1,
  se = 1,
  s_at = 0.375,
  s_nt = 0.375,
  zeta = 0,
  s_eta = 0.01
)
}
\arguments{
\item{nsims}{Number of simulations}

\item{beta}{Treatment effect scaling}

\item{alpha}{Correlation of treatment effect with compliance}

\item{n}{Sample size}

\item{J}{Number of groups}

\item{rde}{Covariance between latency to treat and Y0}

\item{sd}{Variance of latency to treat}

\item{se}{Variance of Y0}

\item{s_at}{Always-taker share}

\item{s_nt}{Never-taker share}

\item{zeta}{Heteroskedasticity}

\item{s_eta}{Quality of prediction of covariates for compliance rate}
}
\value{
Data.frame
}
\description{
Simulate data
}
\examples{
# Create a simulated data.frame
data <- sim.data()
}
