% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/late_rest.R
\name{run.late.rest}
\alias{run.late.rest}
\title{Main function}
\usage{
run.late.rest(
  data,
  yname,
  treat,
  instrument,
  controls,
  n_folds = 2,
  p_th = 0.05,
  joint_est = TRUE,
  csl_est = FALSE
)
}
\arguments{
\item{data}{A data.frame containing the necessary variables to run the model.}

\item{yname}{Name of the dependent variable}

\item{treat}{Name of the treatment variable}

\item{instrument}{Name of the instrument}

\item{controls}{Controls. Either as a formula, eg ~ x1 + x2, or a vector of strings, eg c("x1", "x2")}

\item{n_folds}{Number of folds}

\item{p_th}{Group selection threshold}

\item{joint_est}{Final estimation method}

\item{csl_est}{CS method}
}
\value{
fixest object
}
\description{
Main function
}
\examples{
# Loading packages
library(fixest)

# Generate a simulated dataset
data <- sim.data()

# Run a standard IV
reg1 <- feols(data = data, y ~ 1 | d ~ z)

# Run the test-and-select method
reg2 <- run.late.rest(data = data, yname = "y", treat = "d", instrument = "z", controls = ~g)

# Comparing both regressions
etable(reg1, reg2)
}
